Changed callback calls to support Contao 4.13

+++ dca/tl_files.php
@@ -1,8 +1,5 @@
 <?php
 
-use Hschottm\Tags\DataContainer\FileTableCallbackListener;
-
-
 /**
  * Contao Open Source CMS - tags extension
  *
@@ -11,7 +8,11 @@ use Hschottm\Tags\DataContainer\FileTableCallbackListener;
  * @license LGPL-3.0+
  */
 
-$disabledObjects = deserialize($GLOBALS['TL_CONFIG']['disabledTagObjects'], true);
+if (isset($GLOBALS['TL_CONFIG']['disabledTagObjects'])) {
+	$disabledObjects = deserialize($GLOBALS['TL_CONFIG']['disabledTagObjects'], true);
+} else {
+	$disabledObjects = array();
+}
 if (!in_array('tl_files', $disabledObjects))
 {
 	$GLOBALS['TL_DCA']['tl_files']['palettes']['default'] = str_replace(';meta', ';tags;meta', $GLOBALS['TL_DCA']['tl_files']['palettes']['default']);
@@ -38,25 +39,30 @@ $GLOBALS['TL_DCA']['tl_files']['fields']['tags'] = array
  */
 class tl_files_tags extends \Backend
 {
-	public function removeContentElement($dc)
-	{
-		$this->Database->prepare("DELETE FROM tl_tag WHERE from_table = ? AND tid = ?")
-			->execute($dc->table, $dc->id);
+	public function removeContentElement($source, \DataContainer $dc)
+	{ 
+        $fileModel = \FilesModel::findByPath($source);
+		if ($fileModel != null) {
+			$this->Database->prepare("DELETE FROM tl_tag WHERE from_table = ? AND tid = ?")
+			->execute($dc->table, $fileModel->id);
+		}
 	}
 
-	public function onCopy($source, $destination, \DataContainer $dc)
-	{
-		$objTags = $this->Database->prepare("SELECT * FROM tl_tag WHERE tid = ? AND from_table = ?")
-			->execute($source, $dc->table);
-		$tags = array();
-		while ($objTags->next())
-		{
-			array_push($tags, array("table" => $dc->table, "tag" => $objTags->tag));
-		}
-		foreach ($tags as $entry) {
-			$this->Database->prepare("INSERT INTO tl_tag (tid, tag, from_table) VALUES (?, ?, ?)")
-			->execute($destination, $entry['tag'], $entry['table']);
-		}
-	}
+    public function onCopy($source, $destination, \DataContainer $dc)
+    {
+        $sourceModel = \FilesModel::findByPath($source);
+        $destModel = \FilesModel::findByPath($destination);
+
+        if ($sourceModel != null && $destModel != null) {
+            $objTags = $this->Database->prepare("SELECT * FROM tl_tag WHERE tid = ? AND from_table = ?")->execute($sourceModel->id, $dc->table);
+            $tags = array();
+            while ($objTags->next()) {
+                \array_push($tags, array("table" => $dc->table, "tag" => $objTags->tag));
+            }
+            foreach ($tags as $entry) {
+                $this->Database->prepare("INSERT INTO tl_tag (tid, tag, from_table) VALUES (?, ?, ?)")->execute($destModel->id, $entry['tag'], $entry['table']);
+            }
+        }
+    }
 }
